+++ src/serde/date.rs
@@ -14,6 +14,6 @@ impl TryFrom<Date> for crate::Date {
     type Error = &'static str;
 
     fn try_from(Date(year, ordinal): Date) -> Result<Self, Self::Error> {
-        Self::from_yo(year, ordinal).map_err(|_| "invalid date")
+        Self::from_ordinal_date(year, ordinal).map_err(|_| "invalid date")
     }
 }