+++ tests/error.rs
@@ -3,7 +3,7 @@ use time::{error, Date, Error};
 
 #[cfg(feature = "alloc")]
 fn component_range() -> error::ComponentRange {
-    Date::from_yo(0, 367).unwrap_err()
+    Date::from_ordinal_date(0, 367).unwrap_err()
 }
 
 #[test]